requirements.txt
asgiref==3.6.0
autopep8==2.0.2
black==23.3.0
click==8.1.3
colorama==0.4.6
coverage==7.2.2
Django==4.1.7
django-coverage-plugin==3.0.0
django-environ==0.10.0
django-taggit==3.1.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
flake8==5.0.4
isort==5.12.0
Markdown==3.4.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
psycopg2==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
